修复积分获取接口

+++ mall-product/src/main/java/com/dangdang/bookmall/product/dao/BaseinfoDao.java
@@ -68,7 +68,7 @@ public interface BaseinfoDao extends BaseMapper<BaseinfoEntity> {
      * @return
      */
     @Select("select integral from bmt_baseinfo where id=#{id}")
-    BigDecimal getScoreByIds (int id);
+    Integer getScoreByIds (int id);
 
     @Select("select name from bmt_baseinfo where id=#{id}")
     String getBookNameById(Long id);
@@ -90,6 +90,7 @@ public interface BaseinfoDao extends BaseMapper<BaseinfoEntity> {
     @Select("SELECT name,stock FROM bmt_baseinfo WHERE stock <=5 ORDER BY stock")
 
     List<StockDto> getStockDto();
+
 
 
     @Select("select * from bmt_baseinfo where id=#{id}")
